% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{delete_duplicate_col}
\alias{delete_duplicate_col}
\title{Delete repeated columns in a data frame}
\usage{
delete_duplicate_col(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
The data frame \code{df} excluding all duplicated columns
}
\description{
Delete repeated columns in a data frame
}
\examples{
delete_duplicate_col(df = planaria_cells)
}
