% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{delete_duplicate_row}
\alias{delete_duplicate_row}
\title{Delete repeated rows in a data frame}
\usage{
delete_duplicate_row(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
The data frame \code{df} excluding all duplicated rows
}
\description{
Delete repeated rows in a data frame
}
\examples{
delete_duplicate_row(df = planaria_cells)
}
