% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{export_csv}
\alias{export_csv}
\title{Export data to csv file}
\usage{
export_csv(data, file_path, colnames = TRUE, rownames = TRUE)
}
\arguments{
\item{data}{Content to export to csv file}

\item{file_path}{Name of csv file that is the destination for \code{data}}

\item{colnames}{Boolean indicating whether column names should be
written as first row (default: TRUE)}

\item{rownames}{Boolean indicating whther row names should be
written as first column (default: TRUE)}
}
\description{
Export data to csv file
}
\examples{
data(Fincher_lateral_muscle)
export_csv(data = Fincher_lateral_muscle,
            file_path = "C:/Users/exampleUser/Desktop/planaria.csv")

}
