% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Export data to excel file}
\usage{
export_excel(data, excel_file_path, colnames = TRUE, rownames = TRUE)
}
\arguments{
\item{data}{Content to export to excel file}

\item{colnames}{Boolean indicating whether column names should be written}

\item{rownames}{Boolean indicating whther row names should be written}

\item{excel_file_name}{Name od excel file that is the destination for \code{data}}
}
\description{
Export data to excel file
}
\examples{
export_excel(export = planaria_cells, "planarian.xlsx")
}
