% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{export_txt}
\alias{export_txt}
\title{Export data to txt file}
\usage{
export_txt(data, file_path, colnames = TRUE, rownames = TRUE)
}
\arguments{
\item{data}{Content to export to txt file}

\item{file_path}{Name of txt file that is the destination for \code{data}}

\item{colnames}{Boolean indicating whether column names should be
written as first row (default: TRUE)}

\item{rownames}{Boolean indicating whther row names should be
written as first column (default: TRUE)}
}
\description{
Export data to txt file
}
\examples{
data(Fincher_lateral_muscle)
export_txt(data = Fincher_lateral_muscle,
            file_path = "C:/Users/exampleUser/Desktop/planaria.txt")

}
