% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fincher_muscle_lateral}
\alias{Fincher_muscle_lateral}
\title{Expresssion of genes in lateral muscle cells of the planarian Schmidtea mediterranea.}
\format{A data frame with 26561 rows and 56 variables:
\describe{
  \item{dd_Smed_v4_1000_0_1}{expression of gene, in hits of gene per cell}
  \item{dd_Smed_v4_100_0_1}{expression of gene, in hits of gene per cell}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE111764}
}
\usage{
Fincher_muscle_lateral
}
\description{
A dataset containing the number of hits of each specified gene in over 50
lateral muscle cells.
}
\keyword{datasets}
