% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{import_excel}
\alias{import_excel}
\title{Import data from an excel file}
\usage{
import_excel(file_path, sheet_number, sheet_name = "Sheet1", row_start,
  row_stop, col_start, col_stop, rownames = FALSE, colnames = FALSE)
}
\arguments{
\item{file_path}{Name of excel file}

\item{sheet_number}{Index of sheet in provided excel file}

\item{sheet_name}{Name of the excel sheet to be imported (default: "Sheet1")}

\item{row_start}{Number of first row to import}

\item{row_stop}{Number of last row to iomport}

\item{col_start}{Number of first column to import}

\item{col_stop}{Number of last column to import}

\item{rownames}{Boolean indicating whther first column should become row
names (default: FALSE)}

\item{colnames}{Boolean indicating whether first row should become column
names (default: FALSE)}
}
\description{
Import data from an excel file
}
\examples{
import_excel(file_path = "C:/Users/adamc/Desktop/Fincher_muscle.xlsx",
            sheet_number = 2, sheet_name = "LatMuscle", row_start = 1,
            row_stop = 3250, col_start = 5, col_stop = 55, rownames = TRUE,
            colnames = TRUE)

}
