% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{import_txt}
\alias{import_txt}
\title{Import data from a text file}
\usage{
import_txt(file_path, colnames, rownames)
}
\arguments{
\item{file_path}{Full file path of text file}

\item{colnames}{Boolean indicating whether first row should become column
names (default: FALSE)}

\item{rownames}{Boolean indicating whther first column should become row
names (default: FALSE)}
}
\description{
Import data from a text file
}
\examples{
import_excel(file_path = "C:/Users/adamc/Desktop/Fincher_muscle.xlsx",
            rownames = TRUE, colnames = TRUE)

}
