% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labels.R
\name{label_by_colnames}
\alias{label_by_colnames}
\title{Label two groups of cells based on the contents of their names}
\usage{
label_by_colnames(df, x.start, x.stop, y.start, y.stop, x.name = "group_x",
  y.name = "group_y")
}
\arguments{
\item{df}{A data frame}

\item{x.start}{Index of \code{col.names} at which to start first group of cells}

\item{x.stop}{Index of \code{col.names} at which to stop first group of cells}

\item{y.start}{Index of \code{col.names} at which to start first group of cells}

\item{y.stop}{Index of \code{col.names} at which to stop first group of cells}

\item{x.name}{String label for first cell group (default: "group_x")}

\item{y.name}{String label for second cell group (default: "group_y")}
}
\value{
A named character vector that contains \code{x.name} and \code{y.name}, and all cells that do not qualify as positive or negative are omitted
}
\description{
Label two groups of cells based on the contents of their names
}
\examples{
label_by_colnames(df = Fincher, x.start = 1, x.stop = 11, y.start = 12, y.stop = length(colnames(planaria_cells))
}
