% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{rowselect}
\alias{rowselect}
\title{Select rows from a data frame using a character vector or a numeric vector}
\usage{
rowselect(df, rowchars = c(), rownumbs = c(), complete_names = TRUE,
  df_order = TRUE, print.rownames = FALSE, print.rownumbs = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{rowchars}{A character vector of row names}

\item{rownumbs}{A vector containing the row numbers of the rows to be chosen}

\item{complete_names}{Boolean determining if the entered strings are exact row names (rather than fragments of row names)}

\item{df_order}{Boolean that determines if the rows selected are returned in the order they appear in \code{df} (rather
than the order they appear in \code{rowchars})}

\item{print.rownames}{Boolean determining if selected row names should be printed upon completion}

\item{print.rownumbs}{Boolean determining if selected row numbers should be printed upon completion}
}
\value{
The rows from \code{df} whose names match the character vector \code{rowchars}
}
\description{
Select rows from a data frame using a character vector or a numeric vector
}
\examples{
rowselect(df = planaria_cells, rowchars = c('dd_Smed_v4_12', 'dd_Smed_v4_22'), complete_names = FALSE, print.rownumbs = TRUE)
}
