% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scimone}
\alias{Scimone}
\title{Expresssion of genes in muscle cells of the planarian Schmidtea mediterranea.}
\format{A data frame with 26560 rows and 177 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74360}
}
\usage{
Scimone
}
\description{
A dataset containing the number of hits of each specified gene in over 175
neoblast cells.
}
\keyword{datasets}
