% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{colselect}
\alias{colselect}
\title{Select columns from a data frame using a character vector or numeric vector}
\usage{
colselect(df, colchars = c(), colnumbs = c(), complete_names = TRUE,
  df_order = TRUE, print.colnames = FALSE, print.colnumbs = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{colchars}{A character vector of column names}

\item{complete_names}{Boolean determining if the entered strings are exact column names (rather than fragments of
column names)}

\item{df_order}{Boolean that determines if the columns selected are returned in the order they appear in \code{df} (rather
than the order they appear in \code{colchars})}

\item{print.colnames}{Boolean determining if selected column names should be printed upon completion}

\item{print.colnumbs}{Boolean determining if selected column numbers should be printed upon completion}
}
\value{
A data frame consisting of the columns from \code{df} whose names match the character vector \code{colchars}
in the order that the appear in the column names of \code{df}
}
\description{
Select columns from a data frame using a character vector or numeric vector
}
\examples{
colselect(df = dogs, colchars = c('Pitbu', 'Rottwei'), complete_names = FALSE, df_order = FALSE, print.colnumbs = TRUE)
colselect(df = primaryColors, colnumbs = c(1, 5, 39, 45), complete_names = TRUE, print.colnames = TRUE, print.colnumbs = TRUE)
}
