% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{export_excel}
\alias{export_excel}
\title{Export data to excel file}
\usage{
export_excel(data, file_path, colnames = TRUE, rownames = TRUE,
  name_sheet = "Sheet1", add_to_file = FALSE)
}
\arguments{
\item{data}{Content to export to excel file}

\item{file_path}{Name of excel file that is the destination for \code{data}}

\item{colnames}{Boolean indicating whether column names should be
written as first row (default: TRUE)}

\item{rownames}{Boolean indicating whther row names should be
written as first column (default: TRUE)}

\item{name_sheet}{String indicating what is to be the name of the sheet
containing the newly imported data (default: "Sheet1")}

\item{add_to_file}{Boolean indicating whether data should be added as a
new excel sheet to a current excel file or if data should replace all
existing excel sheets (default: FALSE)}
}
\description{
Export data to excel file
}
\examples{
data(Fincher_lateral_muscle)
export_excel(data = Fincher_lateral_muscle,
            file_path = "C:/Users/exampleUser/Desktop/planaria.xlsx",
            name_sheet = "LatMuscle", add_to_file = TRUE)

}
