% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCDE.R
\name{run_SCDE}
\alias{run_SCDE}
\title{Run single-cell differential expression testing on a specified data frame of cells and genes}
\usage{
run_SCDE(df, labels, file.name, ncores = 1, min_genes_detected = 1000,
  min_reads_per_gene = 5, min_cells_gene_in = 10)
}
\arguments{
\item{df}{A data frame}

\item{labels}{A vector containing labels indicating if certain cells positively or negatively express personal genes}

\item{file.name}{Full file name to use for document containing test results}

\item{ncores}{Number of cores to utilize}

\item{clean.counts_min.lib.size}{Minimum number of genes detected in a cell. Cells with fewer genes will be removed (default: 1.8e3)}
}
\value{
ediff
}
\description{
Run single-cell differential expression testing on a specified data frame of cells and genes
}
\examples{
slit.pos_neg <- pos_neg.label(df = es.mef.small, rowname = 'dd_Smed.v4_12111_0_1')
run_SCDE(df = planaria_cells, labels = slit.pos_neg, file.name = 'slitediff.txt')
}
